change data type of ActivityTotals minutes from int to float

+++ models/activity.go
@@ -14,9 +14,9 @@ type Activity struct {
 }
 
 type ActivityTotals struct {
-	Id         int    `json:"activityId"`
-	Name       string `json:"activityName"`
-	ModuleId   int    `json:"moduleId"`
-	ModuleName string `json:"moduleName"`
-	Minutes    int    `json:"minutes"`
+	Id         int     `json:"activityId"`
+	Name       string  `json:"activityName"`
+	ModuleId   int     `json:"moduleId"`
+	ModuleName string  `json:"moduleName"`
+	Minutes    float32 `json:"minutes"`
 }
